+++ rw-legacy/core.js
@@ -150,16 +150,16 @@ addInputHandler('account_number_splash', function (input) { //acount_number_spla
     var response = input.replace(/\D/g, '');
     if (response == 1) {
         //
-        sayText(msgs('registration_ended', {}, lang));
-        stopRules();
-        // this code is to be uncommented when enrollment starts again
-        // const resumedSession = regSessionManager.resume(contact.phone_number, inputHandlers);
-        // if(!resumedSession){
-        //     var current_menu = msgs('enr_reg_start', {}, lang);
-        //     state.vars.current_menu_str = current_menu;
-        //     sayText(current_menu);
-        //     promptDigits('enr_reg_start', { 'submitOnHash': false, 'maxDigits': max_digits_for_nid, 'timeout': timeout_length });
-        // }
+        // this code is to be uncommented when avocadoe ends
+        // sayText(msgs('registration_ended', {}, lang));
+        // stopRules();
+        const resumedSession = regSessionManager.resume(contact.phone_number, inputHandlers);
+        if(!resumedSession){
+            var current_menu = msgs('enr_reg_start', {}, lang);
+            state.vars.current_menu_str = current_menu;
+            sayText(current_menu);
+            promptDigits('enr_reg_start', { 'submitOnHash': false, 'maxDigits': max_digits_for_nid, 'timeout': timeout_length });
+        }
     } else if(response == 99) { // use 99 to change the language
         lang = service.vars.lang == 'ki' ? 'en' : 'ki'
         service.vars.lang = lang;
@@ -1285,7 +1285,11 @@ addInputHandler('reg_end_ordering_redirect',function(input){
         return null;
     }
     else if(input == 1){
-        
+        // currently ordering is not supported since enrollment for season 22A ended. stopRules.
+        global.sayText(msgs('enr_order_period_finished', {}, lang));
+        global.stopRules();
+        return;
+        // the above two lines should be uncommented when the new season starts
         state.vars.multiple_input_menus = 1;
         var client = get_client(state.vars.account_number, an_pool, true);
         var isDistrictClosed = require('./lib/isDistrictClosed');
@@ -1349,8 +1353,26 @@ addInputHandler('reg_end_ordering_redirect',function(input){
             contact.vars.account_failures = contact.vars.account_failures + 1;
             promptDigits('invalid_input', { 'submitOnHash': false, 'maxDigits': max_digits_for_input, 'timeout': timeout_length })
         }
-    }
-    else{
+    } else if(input == 2) {
+        // start avocado ordering.  this block shoule be commented if the new season starts.
+         var new_client_table = project.initDataTableById(service.vars.rw_reg_client_table_id);
+        var group_details = JSON.parse(state.vars.group_details || '{}')
+        var newClientRow = new_client_table.createRow({vars: {
+            account_number: state.vars.account_number,
+            client_phone_number: state.vars.reg_pn,
+            district: group_details.districtId,
+            first_name: state.vars.reg_name_1,
+            groupId: group_details.groupId,
+            last_name: state.vars.reg_name_2,
+            national_id: state.vars.reg_nid,
+            new_client: '1',
+            registering_phone_number: contact.phone_number,
+            site: group_details.siteId
+        }});
+        newClientRow.save();
+        avocadoTreesOrdering.start(state.vars.account_number,'rw',lang);
+        return;
+    } else{
         sayText(msgs('invalid_input', {}, lang));
         promptDigits('invalid_input', { 'submitOnHash': false, 'maxDigits': max_digits_for_input, 'timeout': timeout_length })
     }